python_manager/requirements.txt
python-can>=4.0.0
PyQt6>=6.0.0